generators: Setup context['filenames'] in Generator.__init__
This ensures that the dict exists so we can use it in
Generator.add_source_path(), etc.

This preserves a single global filenames store because we're assigning
the new dict to the shared context; if a set of generators shares a
context, only the first one will create the fallback dict.